02-build-rag-applications/03-Lab-forked/requirements.txt
llama-index-core==0.11.8
llama-index-llms-huggingface-api
llama-index-embeddings-huggingface
llama-index-readers-web==0.2.2
llama-hub==0.0.79.post1
requests==2.32.2
pydantic==2.10.6

# FIXED VERSIONS
gradio==4.44.0
huggingface_hub
python-dotenv
sentence-transformers
